+++ doc/changelog.rst
@@ -17,6 +17,11 @@ Breaking Changes in 4.0
 - Removed support for Python 2.7, 3.4, and 3.5. Python 3.6+ is now required.
 - Removed :mod:`~pymongo.thread_util`.
 
+Notable improvements
+....................
+
+- Support for MongoDB Versioned API, see :class:`~pymongo.server_api.ServerApi`.
+
 Issues Resolved
 ...............